Update README.md for branch v0.6
Branch v0.6 of this forked repo stores a version update of the CsoundReverb module, to make it compliant with VCV Rack 0.6 API.
While porting the code, I added a few slight changes, so as to fix issues preventing the plugin to run on the Windows platform.
In brief, here are what has been modified:
- general update to VCVRack 0.6 API, as per this issue: https://github.com/VCVRack/Rack/issues/258 
- moved the CsoundReverb.csd file to a separate <csd> folder, and added it to the distributables, as an external asset folder;
- changed the path to load the .csd resource, removing part of the hard-coded path and loading it relatively to the plugin location, via the assetPlugin() method (this fixed the failure in loading the .csd file when running the installed version of Rack under Windows);
- no more need to clone the repo into "Csound" folder (this as a consequence of the above);
- added conditional dllexport attribute implementation, as per this issue: https://github.com/Djack13/VCV-Rack-Csound-Modules/issues/3 
- modified the Makefile adding a Windows conditional block, and including options for the Csound headers and library custom paths to be passed via command line (again, thanks to Rory Walsh for the solution posted in the issue beforehand mentioned).

About the last point, when building from source, one should call <make> passing in the CSOUND_INCLUDE and CSOUND_LIBRARY path parameters. Here an example of <make> command, based on the default installation path of the Csound program under Windows:
$ make CSOUND_INCLUDE="c:/Program Files/Csound6_x64/include" CSOUND_LIBRARY="c:/Program Files/Csound6_x64/lib/"
When not passing arguments, they default to the standard locations highlighted by Djack13 in this issue: https://github.com/Djack13/VCV-Rack-Csound-Modules/issues/2 

Branch v0.6 was tested only under Windows 8.1 x64.
Building with Linux or Mac wasn't tested, but it should work as before.
Testers and contributions are welcome (I could open an issue for that), so to evaluate the chance of opening a PR to the original project.